refactor import paths for constants in challenge edit and create forms
